v.15 add security

+++ FinalProject/src/main/java/com/epam/forum/command/impl/DeletePostByIdCommand.java
@@ -11,6 +11,7 @@ import com.epam.forum.command.Router;
 import com.epam.forum.exception.ErrorTable;
 import com.epam.forum.exception.ServiceException;
 import com.epam.forum.model.entity.Post;
+import com.epam.forum.model.entity.User;
 import com.epam.forum.model.service.PostService;
 import com.epam.forum.validator.DigitValidator;
 
@@ -20,6 +21,7 @@ public class DeletePostByIdCommand implements Command {
 	private static final String ATTRIBUTE_NAME_MESSAGE = "message";
 	private static final String ATTRIBUTE_VALUE_KEY_WRONG_INPUT = "message.wrong.input";
 	private static final String ATTRIBUTE_VALUE_KEY_EMPTY_POST = "message.empty.post";
+	private static final String ATTRIBUTE_NAME_USERNAME = "username";
 
 	private PostService postService;
 
@@ -31,7 +33,8 @@ public class DeletePostByIdCommand implements Command {
 	public Router execute(HttpServletRequest request, HttpServletResponse response) {
 		Router router = new Router();
 		String id = request.getParameter(PARAM_NAME_POST_ID);
-		if (!DigitValidator.isValid(id)) {
+		String username = (String) request.getSession().getAttribute(ATTRIBUTE_NAME_USERNAME);
+		if (username == null || id == null || !DigitValidator.isValid(id)) {
 			request.setAttribute(ATTRIBUTE_NAME_MESSAGE, ATTRIBUTE_VALUE_KEY_WRONG_INPUT);
 			router.setPage(PagePath.TOPIC);
 			return router;
@@ -41,9 +44,18 @@ public class DeletePostByIdCommand implements Command {
 		try {
 			post = postService.findPostById(postId);
 			if (!post.isEmpty()) {
-				postService.delete(post.get());
+				Post deletedPost = post.get();
+				User user = deletedPost.getUser();
+				String usernamePost = user.getUserName();
+				if (usernamePost.equals(username)) {
+					postService.delete(post.get());
+					router.setPage(PagePath.TOPIC);
+				} else {
+					router.setPage(PagePath.FORBIDDEN_PAGE);
+				}
 			} else {
 				request.setAttribute(ATTRIBUTE_NAME_MESSAGE, ATTRIBUTE_VALUE_KEY_EMPTY_POST);
+				router.setPage(PagePath.TOPIC);
 			}
 		} catch (ServiceException e) {
 			logger.error("service exception ", e);
@@ -51,9 +63,8 @@ public class DeletePostByIdCommand implements Command {
 			request.setAttribute(ErrorTable.ERROR_CAUSE, e.getCause());
 			request.setAttribute(ErrorTable.ERROR_LOCATION, request.getRequestURI());
 			request.setAttribute(ErrorTable.ERROR_CODE, HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
-			router.setPage(PagePath.ERROR);	
+			router.setPage(PagePath.ERROR);
 		}
-		router.setPage(PagePath.TOPIC);
 		return router;
 	}
 }
